+++ Cargo.toml
@@ -66,10 +66,12 @@ serde = { version = "=1.0.204", features = ["derive"] }
 serde_json = "=1.0.120"
 serde_with = { version = "=3.9.0", features = ["chrono_0_4"] }
 strum = { version = "=0.26.2", features = ["derive"] }
+struson = { version = "=0.5.0", features = ["serde"] }
 # FIXME: Upgrade once https://github.com/seanmonstar/reqwest/pull/1620 is merged
 reqwest = { git = "https://github.com/thomasqueirozb/reqwest", branch = "base_url", features = [
     "json",
     "stream",
 ] }
+tokio = { version = "=1.39.2", features = ["full"] }
 tracing = { version = "=0.1.40", features = ["attributes"] }
 uuid = { version = "=1.10.0", features = ["v4"], default-features = false }
